requirements-test.txt
-r requirements.txt
django-mock-queries==2.1.5
pytest==5.3.5
pytest-django==3.7.0
flake8==3.8
